# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Streaming de Cursos",
        "description": "API para gerenciamento de cursos, usuários e assinaturas. Para usar a API, faça autenticação pelo endpoint /api/login ou usando OAuth2.",
        "contact": {
            "name": "Suporte API",
            "email": "contato@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/status": {
            "get": {
                "tags": [
                    "Status"
                ],
                "summary": "Verificar o status da API",
                "description": "Retorna o status atual da API e suas conexões",
                "operationId": "getStatus",
                "responses": {
                    "200": {
                        "description": "Status da API",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "API está funcionando!"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-14T22:30:51.860448Z"
                                        },
                                        "environment": {
                                            "type": "string",
                                            "example": "local"
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "12.2.0"
                                        },
                                        "database": {
                                            "type": "string",
                                            "example": "connected"
                                        },
                                        "redis": {
                                            "type": "string",
                                            "example": "connected"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao verificar o status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao verificar serviços"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "oauth2": {
                "type": "oauth2",
                "description": "OAuth2 para autenticação de API",
                "flows": {
                    "password": {
                        "tokenUrl": "http://localhost/oauth/token",
                        "scopes": []
                    }
                }
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}